Flush system.out and add comments narrating the code
The reason system.out.flush() is used is so that all the binary data
gets outputted before the connection is finished. Without this, a chunk
of data is lost at the end of the stream.

(this is the reason why we might get a gray square in the lower
right-hand corner of our output image... we lost the pixels.)